+++ src/compile-library.js
@@ -2,14 +2,13 @@ import mapSharedStyles from './map-shared-styles'
 import replaceSymbols from './replace-symbols'
 import replaceOverrides from './replace-overrides'
 import replaceSharedStyles from './replace-shared-styles'
-
-const {Document} = require('sketch/dom')
+import {Document, Library} from 'sketch/dom'
 
 export default (document, library) => {
   const lookup = mapSharedStyles(document, library)
 
   // replace the symbols
-  const { symbolsMap, docSymbolInstances } = replaceSymbols(document, library)
+  const { symbolsMap, docSymbolInstances } = replaceSymbols(document, library, false)
 
   // replace the styles
   const layerStylesMap = replaceSharedStyles(
@@ -30,9 +29,13 @@ export default (document, library) => {
   })
 
 
-  document.save('/Users/jack.rudelic/Desktop/Monsoy.sketch', {
+  document.save('/Users/jack.rudelic/Desktop/Jack.sketch', {
     saveMode: Document.SaveMode.SaveTo,
   })
+
+  Library.getLibraryForDocumentAtPath(
+    '/Users/jack.rudelic/Desktop/Jack.sketch'
+  )
 
   // reload the inspector to make sure we show the latest changes
   document.sketchObject.reloadInspector()
