--- src/replace-selected-overrides.js
@@ -1,18 +1,0 @@
-export default function (
-  overrideValue,
-  overrides,
-  documentStyles,
-  libraryStyles,
-  key) {
-
-  documentStyles.forEach(function(style){
-    if(style[key] == overrideValue){
-      const styleToImport = libraryStyles[style.name]
-      if(styleToImport){
-        const imported = styleToImport.import()
-        const importedId = imported[key]
-        overrides.value = importedId
-      }
-    }
-  })
-}
